+++ src/OpenTelemetry.Instrumentation.Http/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Recognize the HTTP QUERY method for `http.request.method` (except for .NET 9.0).
-  ([#TODO](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/TODO))
+  ([#3969](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3969))
 
 ## 1.15.0