+++ src/OpenTelemetry.Instrumentation.Http/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Recognize the HTTP QUERY method for `http.request.method` (except for .NET 9.0).
+  ([#3969](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3969))
+
 ## 1.15.0
 
 Released 2026-Jan-21