+++ src/OpenTelemetry.Instrumentation.Http/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 1.15.1
+
+Released 2026-Apr-21
+
 * Recognize the HTTP QUERY method for `http.request.method` (except for .NET 9.0).
   ([#3969](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3969))